Add unit tests for Trufflehog
Create a 'tasks' directory in 'spec' to hold the unit
tests for the different tasks.

Create a 'trufflehog' subdirectory to hold its spec file,
different targets for it to run on, and the expected reports
for those targets. Add a (one-time) shell script to generate those
reports from the targets.